remove 2c2p payment gateway and refactor code

+++ src/CyberSourceSecureAcceptance.php
@@ -4,7 +4,7 @@ namespace Laranex\LaravelMyanmarPayments;
 
 class CyberSourceSecureAcceptance extends CyberSource
 {
-    public function getPaymentData(string $transactionId, string $referenceNumber, int $amount, string $currencyCode = "MMK", string $transactionType = "sale"): array
+    public function getPaymentData(string $transactionId, string $referenceNumber, float $amount, string $currencyCode = "MMK", string $transactionType = "sale"): array
     {
         return $this->processTransaction($transactionId, $referenceNumber, $amount, $currencyCode, $transactionType);
     }
